Move injectables directly under services and remove tests
