+++ apps/authentication-server-ui/src/app/user/overview/actions/delete-user-button/delete-user-button.component.spec.ts
@@ -9,7 +9,7 @@ import {
     runInitializers,
 } from '@dnd-mapp/authentication-server-ui/testing';
 import { provideTranslations } from '../../../../shared';
-import { UsersOverviewStore } from '../../../services/overview/users-overview-store';
+import { UsersOverviewStore } from '../../../services/users-overview-store';
 import { DeleteUserButtonComponent } from './delete-user-button.component';
 
 describe('DeleteUserButtonComponent', () => {
