+++ apps/authentication-server-ui/src/app/user/overview/actions/user-actions.component.spec.ts
@@ -8,7 +8,7 @@ import {
     runInitializers,
 } from '@dnd-mapp/authentication-server-ui/testing';
 import { provideTranslations } from '../../../shared';
-import { UsersOverviewStore } from '../../services/overview/users-overview-store';
+import { UsersOverviewStore } from '../../services/users-overview-store';
 import { UserActionsComponent } from './user-actions.component';
 
 describe('UserActionsComponent', () => {
